+++ src/main/kotlin/dev/aaronhowser/mods/ariadnesthread/registry/ModDataComponents.kt
@@ -4,7 +4,6 @@ import dev.aaronhowser.mods.ariadnesthread.AriadnesThread
 import dev.aaronhowser.mods.ariadnesthread.item.component.BooleanItemComponent
 import dev.aaronhowser.mods.ariadnesthread.item.component.DimensionItemComponent
 import dev.aaronhowser.mods.ariadnesthread.item.component.HistoryItemComponent
-import dev.aaronhowser.mods.ariadnesthread.item.component.LocationItemComponent
 import net.minecraft.core.component.DataComponentType
 import net.neoforged.neoforge.registries.DeferredHolder
 import net.neoforged.neoforge.registries.DeferredRegister
@@ -22,11 +21,6 @@ object ModDataComponents {
     val DIMENSION_COMPONENT: DeferredHolder<DataComponentType<*>, DataComponentType<DimensionItemComponent>> =
         DATA_COMPONENT_REGISTRY.registerComponentType("starting_dimension") {
             it.persistent(DimensionItemComponent.CODEC).networkSynchronized(DimensionItemComponent.STREAM_CODEC)
-        }
-
-    val LOCATION_COMPONENT: DeferredHolder<DataComponentType<*>, DataComponentType<LocationItemComponent>> =
-        DATA_COMPONENT_REGISTRY.registerComponentType("location") {
-            it.persistent(LocationItemComponent.CODEC).networkSynchronized(LocationItemComponent.STREAM_CODEC)
         }
 
     val HISTORY_COMPONENT: DeferredHolder<DataComponentType<*>, DataComponentType<HistoryItemComponent>> =